rename and add a few bad chars for reliability
